refactor: pull items from "ensembl.py" into the __init__.py function
